move in_memory_file_manager.py
