+++ lib/tests/streamlit/streamlit_test.py
@@ -40,9 +40,9 @@ from streamlit.logger import get_logger
 from streamlit.proto.Empty_pb2 import Empty as EmptyProto
 from streamlit.proto.Alert_pb2 import Alert
 
-from streamlit.in_memory_file_manager import _calculate_file_id
-from streamlit.in_memory_file_manager import in_memory_file_manager
-from streamlit.in_memory_file_manager import STATIC_MEDIA_ENDPOINT
+from streamlit.runtime.in_memory_file_manager import _calculate_file_id
+from streamlit.runtime.in_memory_file_manager import in_memory_file_manager
+from streamlit.runtime.in_memory_file_manager import STATIC_MEDIA_ENDPOINT
 
 from tests import testutil
 
@@ -265,7 +265,7 @@ class StreamlitAPITest(testutil.DeltaGeneratorTestCase):
 
     def test_st_audio_options(self):
         """Test st.audio with options."""
-        from streamlit.in_memory_file_manager import _calculate_file_id
+        from streamlit.runtime.in_memory_file_manager import _calculate_file_id
 
         fake_audio_data = "\x11\x22\x33\x44\x55\x66".encode("utf-8")
         st.audio(fake_audio_data, format="audio/mp3", start_time=10)
@@ -873,7 +873,7 @@ class StreamlitAPITest(testutil.DeltaGeneratorTestCase):
     def test_st_video_options(self):
         """Test st.video with options."""
 
-        from streamlit.in_memory_file_manager import _calculate_file_id
+        from streamlit.runtime.in_memory_file_manager import _calculate_file_id
 
         fake_video_data = "\x11\x22\x33\x44\x55\x66".encode("utf-8")
         st.video(fake_video_data, format="video/mp4", start_time=10)
